Extract game screens into components
