+++ src/types.ts
@@ -1,0 +1,7 @@
+import type { ReactNode } from 'react'
+
+export type Game = {
+  title: string
+  description: string
+  component: ReactNode
+}
